+++ libraries/TLC5947/TLC5947.h
@@ -2,13 +2,13 @@
 //
 //    FILE: TLC5947.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.0
+// VERSION: 0.3.0
 //    DATE: 2023-06-17
 // PURPOSE: Arduino library for the TLC5947 24 channel PWM device
 //     URL: https://github.com/RobTillaart/TLC5947
 
 
-#define TLC5947_LIB_VERSION             (F("0.2.0"))
+#define TLC5947_LIB_VERSION             (F("0.3.0"))
 
 
 #include "Arduino.h"
@@ -21,10 +21,14 @@
 class TLC5947
 {
 public:
+  //  single device constructor
   TLC5947(uint8_t clock, uint8_t data, uint8_t latch, uint8_t blank);
+  //  multi device constructor - for daisy chaining)
+  TLC5947(int deviceCount, uint8_t clock, uint8_t data, uint8_t latch, uint8_t blank);
   ~TLC5947();
 
   bool     begin();
+  int      getChannels();
 
   //  fill the internal buffer with PWM values
   //  call write() after all channels are updated (or per channel).
@@ -34,11 +38,13 @@ public:
   uint16_t getPWM(uint8_t channel);
 
   //  percentage wrappers
+  //  also known as duty cycle.
   int      setPercentage(uint8_t channel, float percentage);
   void     setPercentageAll(float percentage);
   float    getPercentage(uint8_t channel);
 
-  //  write the buffer to the TLC5947 device
+  //  write the buffer to the TLC5947 device(s).
+  void     write(int n);
   void     write();
 
   //  control the blank line.
@@ -52,11 +58,13 @@ public:
 
 
 private:
-  uint16_t *_buffer;
-  uint8_t  _clock;
-  uint8_t  _data;
-  uint8_t  _latch;
-  uint8_t  _blank;
+  int        _channels;
+  uint16_t * _buffer;
+  uint8_t    _clock;
+  uint8_t    _data;
+  uint8_t    _latch;
+  uint8_t    _blank;
+  const uint16_t MAXPWM = 4095;
 };